Move the rpcgen lexer to its own file.
